+++ Function/Source/IConnectionPointContainer.cpp
@@ -225,10 +225,10 @@
    char *p,*s = strtok(pszWorkingValues,", ");
    long cntChars = 0;
    while ( s ) {
-      cntChars += strlen(s) + 3;
+      cntChars += (DWORD)strlen(s) + 3;
       theValues.push_back(new double(atof(s)));
-      p = new char[strlen(s) + 1];
-      memset(p,0,strlen(s) + 1);
+      p = new char[(DWORD)strlen(s) + 1];
+      memset(p,0,(DWORD)strlen(s) + 1);
       strcpy(p,s);
       szValues.push_back(p);
       s = strtok(NULL,", ");
@@ -238,16 +238,16 @@
 
    s = pszNames;
    while ( s[0] ) {
-      cntChars += strlen(s) + 2;
+      cntChars += (DWORD)strlen(s) + 2;
       BSTR* bstrName = new BSTR;
-      *bstrName = SysAllocStringLen(NULL,n = strlen(s));
+      *bstrName = SysAllocStringLen(NULL,n = (DWORD)strlen(s));
       MultiByteToWideChar(CP_ACP,0,s,-1,*bstrName,n);
       theNames.push_back(bstrName);
       p = new char[n + 1];
       memset(p,0,n + 1);
       strcpy(p,s);
       szNames.push_back(p);
-      s += strlen(s) + 1;
+      s += (DWORD)strlen(s) + 1;
    }
 
    char *pszResults = new char[cntChars + 1];
@@ -260,14 +260,14 @@
       k++;
    }
    
-   BSTR bstrResults = SysAllocStringLen(NULL,strlen(pszResults) + 1);
-   MultiByteToWideChar(CP_ACP,0,pszResults,-1,bstrResults,strlen(pszResults) + 1);
+   BSTR bstrResults = SysAllocStringLen(NULL,(DWORD)strlen(pszResults) + 1);
+   MultiByteToWideChar(CP_ACP,0,pszResults,-1,bstrResults,(DWORD)strlen(pszResults) + 1);
 
    SAFEARRAYBOUND rgb;
    double *pd;
 
    rgb.lLbound = 1;
-   rgb.cElements = theValues.size();
+   rgb.cElements = (DWORD)theValues.size();
 
    SAFEARRAY* saValues = SafeArrayCreate(VT_R8,1,&rgb);
    SAFEARRAY* saNames = SafeArrayCreate(VT_BSTR,1,&rgb);
@@ -323,7 +323,7 @@
       pvars[1].pparray = &saNames;
    
       pvars[2].vt = VT_I4;
-      pvars[2].lVal = theValues.size();
+      pvars[2].lVal = (DWORD)theValues.size();
    
       pvars[3].vt = VT_I4;
       pvars[3].lVal = iterationNumber;
